Copied twentythirteen theme, to have somthing to work with.
Much of this code will be scratched, and rethough.  However, some parts
of functions.php will be incredibly useful.

I will be converting css to scss files shortly, and the css files will
be included in .gitignore.  You should be able to clone this into your
wordpress themes folder and get something going.

Oh, hopefully we can come up with a quick database that the default
options saved to develop with.

We'll also have to devise a system to keep plugins we're using (and
their settings) in track.