Refactor theme directory
One last refactor of the source directory. This time to keep WordPress theme files (in /src/theme) separate from component code (in /src/components). It should make it clearer these are independent components of the overall solution.

This PR also standardizes the grid — and introduces the first web components to the site! The exa-nameplate replaces the nameplate.php container.

Commit history:

* moving files around
* update style structure
* combine /exa and /hexa functions
* mostly working nameplate
* rollbackable
* polish masthead
* Adding dropdowns
* subtle hover transitions
* fixing the grid
* flexible homepage
* update package